Update all call speaker sinks instead of just active call
This is required to prevent degraded audio if permissions are requested.

If a Chrome user has not given audio device permissions to a site before
a call comes in, the call audio is initiated with the output device ''
as that is the ID Chrome exposes.

The call requires a microphone and requests the user for microphone
permissions. If these are accepted, the computer's full output device
list is available and a `devicechange` browser event is triggered.

This event is solved in the Device and Call classes before the call
is set as the device's `_activeCall` and Call's sinks will not be
updated.

To overcome this, all call sinks can be updated.

There could be a Chrome bug here that is causing this as the degradation
is reverby audio like the track is being played twice. Is Chrome
continuing to play to the (deleted) '' audio device as well as then
switching to play on the 'default' audio device?

MED-706