Evaluate default output device considering `''`
The statements `Array.from(pc.outputs.keys())[0] || 'default'` do not
take into consideration that the output device key could be ''.

On Chrome if no media device permissions are give, the name of the
output device will be ``''`.

When the permissions are later given, the `''` device is never properly
removed because `'' || 'default'` evaluates to `'default'` in
Javascript.

This causes reverby/echoy audio due to audio being played to actually
the same device.

MED-706